Rename ProfileMainScreen2 to ProfileMainScreen
